// packages/storage/src/types.ts
export interface iKeyUser {
  identity: PrivateKey
  /**
   * Auth endpoint used to authenticate this user
   */
  endpoint: string
  /**
   * token is the service token. It can be used to interact with the identity service.
   *
   */
  token: string
  storageAuth?: TextileStorageAuth
}

export interface CreateFolderRequest {
  /**
   * Storage bucket to create the empty folder
   */
  path: string
  /**
   * Path in the bucket to create the empty folder
   */
  bucket: string
}

export interface ListDirectoryRequest {
  /**
   * Path in the bucket to fetch directories from
   */
  path: string
  /**
   * Storage bucket to fetch directory entries
   */
  bucket: string
  /**
   * set recursive to true, if you would like all children of folder entries
   * to be recursively fetched.
   */
  recursive?: boolean
}

/**
 * Represents a member on a shared file
 */
export interface FileMember {
  publicKey: string
  address?: string
  role: PathAccessRole
}

/**
 * Represents an item stored in a storages directory
 */
export interface DirectoryEntry {
  name: string
  path: string
  ipfsHash: string
  isDir: boolean
  sizeInBytes: number
  created: string
  updated: string
  fileExtension: string
  isLocallyAvailable: boolean
  backupCount: number
  members: FileMember[]
  isBackupInProgress: boolean
  isRestoreInProgress: boolean
  uuid: string
  items?: DirectoryEntry[]
  bucket: string
  dbId: string
}

export interface ListDirectoryResponse {
  items: DirectoryEntry[]
}
// ...
export type AddItemDataType =
  | ReadableStream<Uint8Array>
  | ArrayBuffer
  | string
  | Blob

export interface AddItemFile {
  /**
   * path in the bucket where the file should be uploaded.
   * filename would be determined by the last segment in the path
   * so path folder/a_file.txt would have the name `a_file.txt`
   *
   */
  path: string
  /**
   * MimeType of the file being added.
   * This value can be retrieved when opening the file later one.
   *
   */
  mimeType: string
  data: AddItemDataType
  /**
   * progress callback if provided will be called with bytes written to
   * remote while uploading the file.
   *
   */
  progress?: (bytesRead?: number) => void
}

export interface AddItemsRequest {
  bucket: string
  files: AddItemFile[]
}

export interface AddItemsStatus {
  path: string
  status: 'success' | 'error'
  /**
   * Directory entry of uploaded file.
   *
   * Only present if status is 'success'.
   *
   */
  entry?: DirectoryEntry
  error?: Error
}

export interface AddItemsResultSummary {
  bucket: string
  files: AddItemsStatus[]
}

export type AddItemsEventData = AddItemsStatus | AddItemsResultSummary
export type AddItemsEventType = 'data' | 'error' | 'done'
export type AddItemsListener = (data: AddItemsEventData) => void

export interface AddItemsResponse {
  on: (type: AddItemsEventType, listener: AddItemsListener) => void
  /**
   * this function should only be used to listen for the `'done'` event, since the listener would only be called once.
   * or else you could end up having functions leaking (unless you explicitly call the `off()` function).
   */
  once: (type: AddItemsEventType, listener: AddItemsListener) => void
  off: (type: AddItemsEventType, listener: AddItemsListener) => void
}
// ...
export type TxlSubscribeEventData = TxlSubscribeBucketEvent
export type TxlSubscribeEventType = 'data' | 'error' | 'done'
export type TxlSubscribeListener = (data: TxlSubscribeEventData) => void

export interface TxlSubscribeResponse {
  on: (type: TxlSubscribeEventType, listener: TxlSubscribeListener) => void
  /**
   * this function should only be used to listen for the `'done'` event, since the listener would only be called once.
   * or else you could end up having functions leaking (unless you explicitly call the `off()` function).
   */
  once: (type: TxlSubscribeEventType, listener: TxlSubscribeListener) => void
  off: (type: TxlSubscribeEventType, listener: TxlSubscribeListener) => void
}

export interface NotificationSubscribeEvent {
  notification: Notification
  status: 'success' | 'error'
  error?: Error
}

export type NotificationSubscribeEventData = NotificationSubscribeEvent
export type NotificationSubscribeEventType = 'data' | 'error' | 'done'
export type NotificationSubscribeListener = (
  data: NotificationSubscribeEventData
) => void

export interface NotificationSubscribeResponse {
  on: (
    type: NotificationSubscribeEventType,
    listener: NotificationSubscribeListener
  ) => void
  /**
   * this function should only be used to listen for the `'done'` event, since the listener would only be called once.
   * or else you could end up having functions leaking (unless you explicitly call the `off()` function).
   */
  once: (
    type: NotificationSubscribeEventType,
    listener: NotificationSubscribeListener
  ) => void
  off: (
    type: NotificationSubscribeEventType,
    listener: NotificationSubscribeListener
  ) => void
}

/**
 * FullPath represents full path information to a file.
 * `dbId` is optional and only required for when re-sharing files in another db.
 */
export interface FullPath {
  path: string
  bucket: string
  bucketKey?: string
  dbId?: string
  uuid?: string
}

/**
 * InvitationStatus represents the different statuses a file invitation could have
 */
export enum InvitationStatus {
  PENDING = 0,
  ACCEPTED,
  REJECTED,
}

/**
 * Invitation represents a file invitation
 * `invitationID` is the same as the underlying message ID from Textile
 */
export interface Invitation {
  inviterPublicKey: string
  inviteePublicKey: string
  invitationID?: string
  status: InvitationStatus
  itemPaths: FullPath[]
  keys: string[]
}
// ...
export enum NotificationType {
  UNKNOWN = 0,
  INVITATION = 1,
  USAGEALERT = 2,
  INVITATION_REPLY = 3,
  REVOKED_INVITATION = 4,
}

export interface Notification {
  id: string
  from: string
  to: string
  body: Uint8Array
  decryptedBody: Uint8Array
  type: NotificationType
  createdAt: number
  readAt?: number
  relatedObject?: Invitation
}